refactor: make plugin manager more generic
