+++ server/algorithms/src/lib.rs
@@ -2,6 +2,7 @@ use model;
 
 pub mod bootstrap;
 pub mod clustering;
+mod plugin;
 mod project;
 pub mod routing;
 mod rtree;
